requirements.txt
librosa
torch
torchvision
scipy
matplotlib
numpy
pandas
tensorflow
tensorflow-probability
torchmetrics
seaborn
ipykernel
scaper
wget
Albumentations
opendatasets
kaggle
pillow
Image